% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ef.R
\name{find_ef}
\alias{find_ef}
\title{Function to return emission factors}
\usage{
find_ef(name)
}
\arguments{
\item{name}{character; Word to be searched among the names of the cpt gradients.}
}
\value{
names that satisfy the search.
}
\description{
\code{\link{find_ef}} returns the emission factors data.frame with the parameters
searched.
}
\note{
This functions runs grep.
}
\examples{
\dontrun{
#do not run
}
}
